Wire apollo / artemis / optsync over libmempool-rs
Replicates the synchs-pilot shape to the remaining three protocols:

  - Each Context owns `batch_store: RocksStore`, a
    `pending_batches: VecDeque<BatchHash>` queue fed by
    `rx_mem_to_consensus`, and a `tx_consensus_to_mem` handle for
    future synchronizer plumbing.
  - Proposal wire: `NewProposal(p, b, batch)` for apollo/synchs/optsync
    (leader inlines the batch so followers don't need a sync
    round-trip). Apollo `Response` + artemis `NewBlock` / `Response`
    picked up the same shape.
  - On incoming proposal/response, receivers hash-check the batch and
    persist it into rocksdb via `cx.persist_batch` before voting /
    delivering.
  - On commit, receivers read the batch back from rocksdb and hydrate
    per-tx hashes into `ClientMsg::NewBlock`. Apollo's special-client
    path hydrates at propose time instead (leader has the batch in
    hand already).
  - Client-side latency tracking moves from `block.body.tx_hashes`
    (which no longer exists) to the server-hydrated hashes that now
    ride in `ClientMsg::NewBlock`. apollo/artemis clients grow a
    `tx_hash_map` so the round/chain-based commit loop can match
    hashes when the commit round has moved on.
  - Node binaries spawn the full libmempool-rs pipeline: rocksdb
    store, `Batcher` w/ `CountSealer(block_size)`, `Mempool::spawn`
    with TcpSimpleSender for peer sync, plus the libnet-rs TLS
    transports already in place for consensus.
  - Client binaries swap `TlsReliableSender<Replica, Transaction>`
    for `TcpSimpleSender` (libmempool's client endpoint is plain
    TCP). Commit-notification path stays on TLS.

Stress-test sweep (50k tx, blk=400, payload=0):

  Apollo              3   1   19680 tx/s     528 ms
  Apollo              7   3    7949 tx/s     920 ms
  Artemis             3   1   27815 tx/s     357 ms
  Artemis             7   3   18569 tx/s     726 ms
  Sync HotStuff       3   1   69747 tx/s     127 ms
  Sync HotStuff       7   3   44721 tx/s     207 ms
  Opt Sync            3   1  169132 tx/s      51 ms
  Opt Sync            7   3   38132 tx/s     248 ms

Apollo's round-chain client reconstruction is a bottleneck given the
new hash-lookup via tx_hash_map; to investigate in a follow-up.
Sync HotStuff and Opt Sync gain 3-5x from the mempool offload.

+++ consensus/artemis/src/node/commit.rs
@@ -8,10 +8,10 @@ use super::*;
 /// Caller needs to ensure that `cx.round > cx.num_faults()`
 pub fn do_commit(cx: &mut Context) {
     log::debug!("Trying to commit");
-    debug_assert!(cx.round() > cx.num_faults());
+    debug_assert!(cx.round() > cx.num_faults() as u64);
 
     // Get the r-f^th vote
-    let commit_round = cx.round() - cx.num_faults();
+    let commit_round = cx.round() - cx.num_faults() as u64;
     let v = cx.vote_chain.get(&commit_round).unwrap();    
 
     let mut com_hash = v.hash.clone();
